+++ meaningless/__init__.py
@@ -10,6 +10,8 @@ from meaningless.bible_json_downloader import JSONDownloader
 from meaningless.bible_json_extractor import JSONExtractor
 from meaningless.bible_xml_downloader import XMLDownloader
 from meaningless.bible_xml_extractor import XMLExtractor
+from meaningless.bible_csv_downloader import CSVDownloader
+from meaningless.bible_csv_extractor import CSVExtractor
 from meaningless.bible_web_extractor import WebExtractor
 # Ignore the base error class, but include all the other exception types
 from meaningless.utilities.exceptions import (
@@ -20,4 +22,4 @@ from meaningless.utilities.exceptions import (
 )
 # Include the file interfaces, mainly as an out-of-the-box mechanism for reading downloaded files
 # as well as writing output using the information obtained from the extractors.
-from meaningless.utilities import yaml_file_interface, json_file_interface, xml_file_interface
+from meaningless.utilities import yaml_file_interface, json_file_interface, xml_file_interface, csv_file_interface