+++ test/system_tests_bible_translations.py
@@ -268,6 +268,11 @@ class UnitTests(unittest.TestCase):
         self.check_baseline_passages(translation)
         self.check_omitted_passages(translation)
 
+    def test_translation_nivuk(self):
+        translation = 'NIVUK'
+        self.check_baseline_passages(translation)
+        self.check_omitted_passages(translation)
+
 
 if __name__ == "__main__":
     unittest.main()