+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.5.0
+- Added CSV Extractor and CSV Downloader
+- Added csv_file_interface module to assist with CSV file access
+- Added translation support for: NIVUK
+- Drop library support for Python 3.6
+
 ## 0.4.0
 - Documentation is now publicly available through GitHub Pages. See https://daniel-tran.github.io/meaningless/
   - This was previously an undocumented change as part of the 0.3.0 release.